+++ tests/004-heap-alloc.c
@@ -8,14 +8,14 @@ int main(void) {
 
     uint32_t *p1 = mp_alloc(&alloc, sizeof(*p1));
     *p1          = 67;
-    expect_eq(*p1, 67, "%u");
+    expect_eq(*p1, 67u, "%u");
 
     uint64_t *p2 = mp_realloc(&alloc, p1, sizeof(*p1), sizeof(*p2));
     *p2          = 69;
-    expect_eq(*p2, (uint64_t) 69, "%zu");
+    expect_eq(*p2, 69uL, "%zu");
 
     uint64_t *p3 = mp_dup(&alloc, p2, sizeof(*p2));
-    expect_eq(*p3, (uint64_t) 69, "%zu");
+    expect_eq(*p3, 69uL, "%zu");
     mp_free(&alloc, p2, sizeof(*p2));
     mp_free(&alloc, p3, sizeof(*p3));
